pyproject.toml
[project]
name = "worksheet-maker-gui"
version = "1.0.0"
description = "React and FastAPI learning material maker for word search, word flicker, worksheets, and Dobble cards."
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
  "fastapi>=0.136.1",
  "httpx>=0.28.1",
  "pillow>=12.2.0",
  "python-docx>=1.2.0",
  "python-pptx>=1.0.2",
  "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
  "mypy>=1.20.2",
  "pytest>=9.0.3",
  "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [
  "basic_gui.py",
  "card_maker.py",
  "dobble_maker.py",
  "label_maker.py",
  "learning_material_maker.py",
  "word_flicker_maker.py",
  "word_worksheet_maker.py",
  "wordsearch_maker.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["backend"]
